fix(core): center y axis relative to data rectangle
Currently the y axis title is centered against the dimensions of the entire chart. This can be disorienting when a bar chart has an odd number of bars, because one of the bars is clearly in the center of the data space but it appears to be off-center relative to the axis title, which is including the margin for the other axis in its positioning calculation.