feat(core): encoding titles
The encoding property can be applied to an encoding directly instead of to an axis object. This is logically appropriate for encodings that don't imply a linear axis, such as color and size. Once it becomes a valid part of the specification, however, you have to listen to it! If an encoding-level title is present, use it when rendering text attached to the axis or legend. The axis-level or legend-level title takes priority if both are present, since it's more specific.