chore(core): change dimensions argument name
The variable name "panelDimensions" is a vestigial term that used to be necessary to disambiguate dimensions that include the legend from graphic-only dimensions when this was all running inside the chart() function. It's no longer necessary now that it has its own module and its own function, and changing it to "explicitDimensions" makes it clearer that it's optional, and explains what the functionality is when it is provided.