chore(core): add parentheses to function name in detach error
Add parentheses to the function name in the error message. This can be the difference between a message like "could not run function renderer" versus "could not run function renderer()" – the former seems like it might be trying to render a function, whatever that would mean, while the latter makes it clearer that "renderer" is the name of the function that failed.