chore(core): scale extension granularity
Bar length and bar starting position are logically connected, but the idea behind using a structured process for scale extension was that they could in theory be applied individually. In other words, prior to this change, the scale extension for length was actually operating as a flag that would trigger scale extension for both bar starting position and bar length. This can be cleaned up. In practice they'll still always be used together.