feat(core): handle negative values in x encoding
Up until now, the only shenanigans necessary have been with the y axis, which needs to be inverted from the SVG coordinate system in order to make sense for data values. The x axis doesn't require that sort of inversion, so it was okay to just count up from zero to get the bar length or starting position. In order to handle nonzero baselines along the x axis, though, it's necessary to do a little more juggling.