fix(docs): remove invalid argument types
When building the documentation, JSDoc doesn't like the version of this function header that includes input types for the returned function. Removing them is a bit less type safe, but it's probably not a big deal, and it prevents errors from being logged to the terminal when documentation is generated.