chore(core): generalize axis offset function
The axis offset function is still hard coded to only handle charts where y is the quantitative encoding, which manifests as a misplaced y axis for bar charts where the x axis is quantitative and there are negative values in the chart.

This is because the axis offset function is actually hard-coded to only handle that orientation! So the first step is to change the name and interface of that function into something that could theoretically handle the same position computation for the x axis. This doesn't actually fix the problem, hence the error handling, but now the solution can be scoped to the offset function.