chore(core): describe optional new value for nesting function
The nesting function can be used either as a getter or a setter, but currently is only used as a getter. Adding documentation for the unused third parameter appeases the linter.

This is restricted to a couple of types to avoid the use of an {any} type, but it's just a setter for arbitrary values, so that can be expanded as necessary when real-world cases emerge in the future.