fix(core): alternate range index for line marks
Since the SVG coordinate system runs from top to bottom and this is typically hidden by inverting the scale range, extracting the endpoint length for the grid line requires checking the mark type.

Specifically, this manifests as a fix for axis ticks along a continuous x axis for line charts which have been inverted to run from top to bottom instead of left to right. It's definitely an edge case!