fix(core): grid lines for negative values
Selecting one of the range endpoints and using it as the length of the grid lines ignores the case of negative values where the largest integer represents only the portion of the data rectangle above the zero baseline. In order to handle that case, the length should be set to the total number of pixels separating the two range endpoints, not just the largest integer in cases where the other range endpoint is zero.