chore(core): use bar width when assigning stack encoders
The stackEncoders() function handles reassigning various abstract data-level concepts into visual space. For example, the bar "length" can exist in any orientation, but for a vertical bar chart it needs to be reassigned to the y dimension and for a horizontal bar chart it needs to be reassigned to the x dimension.

For the new stricter Cartesian layout code to work properly with single-bar charts – one bar proportionally divided into several colors, sort of like a linear version of a pie chart – it needs to reference the barWidth() function. For cases where we're not actually dealing explicitly with a bar mark, it continues to use the more generic step() function.