chore(core): set width for single bar charts
Up until now the barWidth() function has been identical to the step() function – literally just a reassignment of the variable containing the function in order to provide a semantically appropriate name under which to call it depending on whether the chart is a bar chart or not. Now they need to diverge, because barWidth() needs to be able to provide a width for a bar chart that only consists of a single bar.