fix(core): center x axis label
Somewhere along the way this position logic became more convoluted than it needs to be. The result was an x axis text label that is sometimes positioned slightly off-center relative to the mark content. This is visible when the number of axis ticks or nominal bands on the axis is both small and odd – it's easy to see that there's a disagreement about what constitutes the "center" when there are only three bars, for example.

The underlying cause for this is an accidental philosophical disagreement about what dimensions should mean. The original Vega Lite renderer uses them to define the dimensions of the marks area, and adds extra things like axes or titles on top of that. In contrast, bisonica has always interpreted dimensions instructions as describing the outer bounds of the node, the total size of the chart.

One could argue that this is an incompatibility between the two renderers, but it may also be fair to say that bisonica chose the interpretation that is more useful in the real world.