feat(core): aria label for legend title and content
Generate a unique identifier for the legend title, then refer to that with aria-labelledby to create a more explicit association between the legend content and the title.

Looking up the header content from inside the swatch function is a little fiddly, but it's cleaner than changing the argument signature, and it's enforced through a test.