chore(core): descriptive error for misconfigured title
One common stumbling block when getting started with bisonca is that unlike Vega Lite it requires every specification to have a descriptive title that can be rendered for screen readers, and will refuse to render a chart at all if that is omitted. There's already a specific error message for that common misconfiguration, but a secondary point of confusion is that per the Vega Lite documentation, specification.title is not simply a text string – it's an object which has a text property. This adds an even more descriptive error message to be thrown in cases where the title is present but it is a string instead of an object.

https://vega.github.io/vega-lite/docs/title.html