feat(core): type check specifications
Until now, the type checker has only verified that a specification is an object. This is as loose a definition as is possible in JavaScript! It can be stricter.

The type definition only looks for top-level properties, all of which are objects, and most of which are optional. This is enough to disambiguate Vega Lite from other objects without fully submitting to the tyranny of aggressive type checking.