chore(core): add typedef for layer specifications
Now that we have a JSDoc @typedef for Vega Lite specifications which distinguish them from normal objects, we can also implement the same thing for the layerSpecification which is generated for each layer when rendering with view composition.

The layerSpecifications are constructed loosely and only ever used within this one module – which is why the typedef is located here instead of in types.d.js – and consequently I also didn't want the type to be particularly specific, because fighting against the type system while working on something as abstract as view composition sounds like a nightmare. Accordingly this type is just an alias for {object}, but using the alias lets us more clearly outline types for inputs and outputs of the functions, even if the underlying types themselves are still only vaguely defined.