chore(tooling): don't require properties in layerSpecification typedef
The JSDoc plugin for eslint is configured by default to complain about {object} typedefs that have no properties defined. In most cases this is what you want, but here I want a loosely defined type that only serves as a way of signaling the flow of an argument through various functions without over-specifying what the type is – routing, but not shape.