fix(core): disable impossible downloads
Extend the feature test for downloads to check for the presence of the URL global, which is used to package up the data into a file. If this doesn't exist, don't render the menu item for downloads at all. Previously the menu item would render but the download would fail.